add comment to some servlet, fixed bucket show bugs

+++ src/main/java/com/z/shop/servlet/RemoveFromBucketServlet.java
@@ -24,27 +24,40 @@ public class RemoveFromBucketServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         HttpSession session = request.getSession();
 
+        /**
+         * Deleting from bucket reserved but not bought product.
+         * User can delete only from own bucket, from current session and DB.
+         * Admin can delete from any user bucket from DB.
+         * */
+
+
         Integer productId = Integer.valueOf(request.getParameter("productId"));
         List<Bucket> buckets = (List<Bucket>) session.getAttribute("buckets");
 
 
         User user = (User) session.getAttribute("user");
-        System.out.println(user);
+
         String bucketIdFromRequest = request.getParameter("bucketId");
-        System.out.println(bucketIdFromRequest);
+
         if(bucketIdFromRequest != null && !bucketIdFromRequest.isEmpty()){
-            System.out.println("inside == null");
+
             Integer bucketId = Integer.valueOf(bucketIdFromRequest);
-            System.out.println( user.getRole().toString().equals("ADMIN"));
+
             if (user != null && user.getRole().toString().equals("ADMIN")){
                 bucketService.delete(bucketId);
-                System.out.println("Admin bucket");
+
                 response.sendRedirect("/showAllBuckets");
             }
             if (user != null && user.getRole().toString().equals("USER")){
                 Bucket bucket = bucketService.read(bucketId);
-                if (bucket.getUserId().equals(user.getId())) {
-                    bucketService.delete(bucketId);
+                if (bucket != null) {
+                    if (bucket.getUserId().equals(user.getId())) {
+                        bucketService.delete(bucketId);
+                        buckets = buckets.stream().filter(buck -> buck.getId().intValue() != bucketId.intValue()).collect(Collectors.toList());
+                        session.setAttribute("buckets", buckets);
+                        response.sendRedirect("/bucket");
+                    }
+                }else{
                     buckets = buckets.stream().filter(buck -> buck.getId().intValue() != bucketId.intValue()).collect(Collectors.toList());
                     session.setAttribute("buckets", buckets);
                     response.sendRedirect("/bucket");
